Fix consecutive iterations of base_loop
Before, we were not reusing the guards already in
intermediate_brightness_states, and were instead failing immediately
after finding a None in optional_guards.
Now, optional_guards is a list of guards that are NOT already in
intermediate_brightness_states. We then go through and add each guard to
intermediate_brightness_states as a partially initialized MonitorInfo,
and then fully initialize everything in the next loop